+++ gmd/Common/RepoState.cs
@@ -1,10 +1,19 @@
+using gmd.Common.Private;
+
 namespace gmd.Common;
+
+public class BranchOrder
+{
+    public string Branch { get; set; } = "";
+    public string Other { get; set; } = "";
+    public int Order { get; set; } = 0;
+}
 
 public class RepoState
 {
     public List<string> Branches { get; set; } = new List<string>();
     public Dictionary<string, int> BranchColors { get; set; } = new Dictionary<string, int>();
-
+    public List<BranchOrder> BranchOrders { get; set; } = new List<BranchOrder>();
 }
 
 interface IRepoState