+++ gmd/Common/State.cs
@@ -1,3 +1,4 @@
+using gmd.Common.Private;
 using static System.Environment;
 
 namespace gmd.Common;
@@ -30,8 +31,8 @@ public class State
 {
     public List<string> RecentFolders { get; set; } = new List<string>();
     public List<string> RecentParentFolders { get; set; } = new List<string>();
-
     public Releases Releases { get; set; } = new Releases();
+    public string GitVersion { get; set; } = "";
 }
 
 interface IState