+++ gmd/Cui/AboutDlg.cs
@@ -22,24 +22,21 @@ class AboutDlg : IAboutDlg
 
     public void Show()
     {
-        UI.Post(async () =>
-        {
-            var releases = states.Get().Releases;
-            var typeText = releases.IsPreview ? "(preview)" : "";
-            var gmdVersion = Build.Version();
-            var gmdBuildTime = Build.Time().Iso();
-            var gmdSha = Build.Sha();
-            var isAvailable = Build.Version() < Version.Parse(releases.LatestVersion);
-            if (!Try(out var gitVersion, out var e, await git.Version())) gitVersion = "0.0";
+        var releases = states.Get().Releases;
+        var typeText = releases.IsPreview ? "(preview)" : "";
+        var gmdVersion = Build.Version();
+        var gmdBuildTime = Build.Time().Iso();
+        var gmdSha = Build.Sha();
+        var isAvailable = Build.Version() < Version.Parse(releases.LatestVersion);
+        var gitVersion = states.Get().GitVersion;
 
-            var msg =
-                $"Version:   {gmdVersion} ({gmdSha}) \n" +
-                $"Built:     {gmdBuildTime}\n" +
-                (isAvailable ? $"Available: {releases.LatestVersion} {typeText}\n" : "") +
-                $"Git:       {gitVersion} ";
+        var msg =
+            $"Version:   {gmdVersion} ({gmdSha}) \n" +
+            $"Built:     {gmdBuildTime}\n" +
+            (isAvailable ? $"Available: {releases.LatestVersion} {typeText}\n" : "") +
+            $"Git:       {gitVersion} ";
 
-            UI.InfoMessage("About", msg);
-        });
+        UI.InfoMessage("About", msg);
     }
 }