+++ gmd/Cui/RepoViewMenus.cs
@@ -161,14 +161,67 @@ class RepoViewMenus : IRepoViewMenus
             SubMenu("Open/Clone Repo", "", GetOpenRepoItems()),
             Item("Change Branch Color", "G", () => cmds.ChangeBranchColor(), () => !repo.Branch(repo.RowCommit.BranchName).IsMainBranch),
             SubMenu("Set Branch", "", GetSetBranchItems(), () => GetSetBranchItems().Any()),
+            SubMenu("Move Branch left/right", "", GetMoveBranchItems(), () => GetMoveBranchItems().Any()),
             Item("Help ...", "H", () => cmds.ShowHelp()),
             Item("Config ...", "", () => configDlg.Show(repo.RepoPath)),
             Item("About ...", "", () => cmds.ShowAbout())
         );
     }
 
+    IEnumerable<MenuItem> GetMoveBranchItems()
+    {
+        var items = Enumerable.Empty<MenuItem>();
 
-    private IEnumerable<MenuItem> GetStashMenuItems()
+        // Get possible local, remote, pull merge branches of the row branch
+        var rowCommonName = repo.RowBranch.CommonName;
+        var rowBranches = repo.Branches.Where(b => b.CommonName == rowCommonName);
+
+        // Get all branches that overlap with any of the row branches
+        var overlappingBranches = rowBranches
+            .SelectMany(b => repo.Graph.GetOverlappinBranches(b.Name))
+            .Distinct()
+            .ToList();
+
+        if (!overlappingBranches.Any()) return items;
+
+        // Sort on left to right shown order
+        Sorter.Sort(overlappingBranches, (b1, b2) => b1.X < b2.X ? -1 : b1.X > b2.X ? 1 : 0);
+
+        // Find possible branch on left side to move to before (skip if ancestor)
+        Branch? leftBranch = null;
+        for (int i = 0; i < overlappingBranches.Count; i++)
+        {
+            var b = overlappingBranches[i];
+            if (b.B.CommonName == rowCommonName) break;
+            leftBranch = b.B;
+        }
+        var leftCommonName = leftBranch != null && !IsAncestor(leftBranch, repo.RowBranch) ? leftBranch.CommonName : "";
+
+        // Find possible branch on right side to move to after (skip if ancestor)
+        Branch? rightBranch = null;
+        for (int i = overlappingBranches.Count - 1; i >= 0; i--)
+        {
+            var b = overlappingBranches[i];
+            if (b.B.CommonName == rowCommonName) break;
+            rightBranch = b.B;
+        }
+        var rightCommonName = rightBranch != null && !IsAncestor(repo.RowBranch, rightBranch) ? rightBranch.CommonName : "";
+
+        // Add menu items if movable branches found
+        if (leftCommonName != "")
+        {
+            items = items.Append(Item($"<= (Move {repo.RowBranch.DisplayName} left of {leftBranch!.DisplayName})", "", () => cmds.MoveBranch(repo.RowBranch.CommonName, leftCommonName, -1)));
+        }
+        if (rightCommonName != "")
+        {
+            items = items.Append(Item($"=> (Move {repo.RowBranch.DisplayName} right of {rightBranch!.DisplayName})", "", () => cmds.MoveBranch(repo.RowBranch.CommonName, rightCommonName, +1)));
+        }
+
+        return items;
+    }
+
+
+    IEnumerable<MenuItem> GetStashMenuItems()
     {
         return EnumerableEx.From(
             Item("Stash Changes", "", () => cmds.Stash(), () => !repo.Status.IsOk),
@@ -272,6 +325,12 @@ class RepoViewMenus : IRepoViewMenus
         items.AddRange(repo.Branches
             .Where(b => !b.IsRemote && !b.IsCurrent && b.HasLocalOnly && !b.HasRemoteOnly)
             .Select(b => (Item($"Push {ToBranchMenuName(b)}", "", () => cmds.PushBranch(b.Name)))));
+
+        if (repo.CurrentBranch != null && repo.CurrentBranch.RemoteName == "")
+        {
+            items.Add(Item($"Publish {repo.CurrentBranch.DisplayName}", "",
+            () => cmds.PublishCurrentBranch()));
+        }
 
         return items.DistinctBy(b => b.Title);
     }
@@ -419,12 +478,10 @@ class RepoViewMenus : IRepoViewMenus
 
         var liveBranches = allBranches
             .Where(b => b.IsGitBranch)
-            .DistinctBy(b => b.CommonName)
-            .OrderBy(b => b.CommonName);
+            .OrderBy(b => b.DisplayName);
 
         var liveAndDeletedBranches = allBranches
-            .DistinctBy(b => b.CommonName)
-            .OrderBy(b => b.CommonName);
+            .OrderBy(b => b.DisplayName);
 
         var recentBranches = liveAndDeletedBranches
             .OrderBy(b => repo.Repo.AugmentedRepo.CommitById[b.TipId].GitIndex)
@@ -432,7 +489,7 @@ class RepoViewMenus : IRepoViewMenus
 
         var ambiguousBranches = allBranches
             .Where(b => b.AmbiguousTipId != "")
-            .OrderBy(b => b.CommonName);
+            .OrderBy(b => b.DisplayName);
 
         var items = EnumerableEx.From(
             SubMenu("Recent", "", ToShowHiarchicalBranchesItems(recentBranches)),
@@ -478,7 +535,8 @@ class RepoViewMenus : IRepoViewMenus
         // Group by first part of the b.commonName (if '/' exists in name)
         var groups = branches
             .GroupBy(b => b.CommonName.Split('/')[0])
-            .OrderBy(g => g.Key);
+            .OrderBy(g => g.Key)
+            .OrderBy(g => g.Count() > 1 ? 0 : 1);  // Sort groups first;
 
         // If only one item in group, then just show branch, otherwise show submenu
         return ToMaxBranchesItems(groups.Select(g =>
@@ -556,6 +614,21 @@ class RepoViewMenus : IRepoViewMenus
 
         return name.Replace('_', '-');
     }
+
+    bool IsAncestor(Branch b1, Branch? b2)
+    {
+        while (b2 != null)
+        {
+            if (b2.CommonName == b1.CommonName)
+            {
+                return true;
+            }
+            b2 = repo.Branches.FirstOrDefault(b => b.Name == b2.ParentBranchName);
+        }
+
+        return false;
+    }
+
 
     string Sid(string id) => id == Repo.UncommittedId ? "uncommitted" : id.Sid();
 }