+++ gmd/Server/Private/Augmented/Repo.cs
@@ -78,6 +78,7 @@ public record Branch(
     string LocalName,
 
     string ParentBranchName,
+    string ParentBranchCommonName,
     string PullMergeBranchName,
 
     bool IsGitBranch,